move login fetch functionality to userActions as loginUser
