+++ src/components/Home.js
@@ -5,7 +5,7 @@ import { useState, useEffect } from 'react'
 
 const url = 'http://localhost:3001/api/v1/users/show'
 
-const Home = ({ logoutUser }) => {
+const Home = ({ user: { email, username }, loading, logoutUser }) => {
     const [likesCount, setLikesCount] = useState(0)
     const [commentsCount, setCommentsCount] = useState(0)
 
@@ -27,15 +27,16 @@ const Home = ({ logoutUser }) => {
 
     return(
         <div>
+            { loading ? "loading" : ""}
             <Page narrowWidth>
                 <br></br>
                 <Card>
                     <br></br>
                     <DisplayText size="Medium">Username</DisplayText>
-                    {/* <DisplayText size="Small">{user.username}</DisplayText> */}
+                    <DisplayText size="Small">{username}</DisplayText>
                         <br></br>
                     <DisplayText size="Medium">Email</DisplayText>
-                    {/* <DisplayText size="Small">{user.email}</DisplayText> */}
+                    <DisplayText size="Small">{email}</DisplayText>
                         <br></br>
                     <Button size="large" outline={true} primary={true} onClick={logoutUser}>Logout</Button>
                         <br></br>
@@ -46,4 +47,4 @@ const Home = ({ logoutUser }) => {
     )
 }
 
-export default connect(({ user }) => ({ user }), { logoutUser })(Home)
+export default connect(({ user, loading }) => ({ user, loading }), { logoutUser })(Home)
