+++ src/components/LoginForm.js
@@ -1,12 +1,12 @@
 import { Form, FormLayout, TextField, Button, Page, Card, DisplayText } from '@shopify/polaris'
 import { useState } from 'react'
 import { connect } from 'react-redux'
-import { setUser } from '../actions/userActions'
+import { setUser, loginUser } from '../actions/userActions'
 import { useHistory } from 'react-router-dom'
 
 const base_url = "http://localhost:3001/api/v1/login"
 
-const LoginForm = ({ setUser }) => {
+const LoginForm = ({ setUser, loginUser }) => {
     let history = useHistory();
     const [email, setEmail] = useState("");
     const [password, setPassword] = useState("");
@@ -16,31 +16,12 @@ const LoginForm = ({ setUser }) => {
     const handlePasswordChange = e => setPassword(e.target.value);
 
     const handleSubmit = e => {
-        e.preventDefault()
-
+        e.preventDefault();
         const user = {
             email,
             password
         }
-        console.log(user)
-
-        const configObj = {
-            method: "POST",
-            headers: {
-                "Content-Type": "application/json",
-                "Accepts": "application/json"
-            },
-            body: JSON.stringify(user)
-        }
-
-        fetch(base_url, configObj)
-        .then(resp => resp.json())
-        .then(returnedUserData => {
-            localStorage.setItem("token", returnedUserData.jwt)
-            setUser(returnedUserData)
-            history.push('/explore')
-        })
-
+        loginUser(user, history);
         setEmail("");
         setPassword("");
     }
@@ -78,4 +59,4 @@ const LoginForm = ({ setUser }) => {
     )
 }
 
-export default connect(null, { setUser })(LoginForm)
+export default connect(null, { setUser, loginUser })(LoginForm)
